More refactoring of existing leetcode solns
