--- leetcode/Makefile
@@ -1,10 +1,0 @@
-build:
-	go build -v ./...
-
-test:
-	go test -v -coverprofile=coverage.out ./...
-	go tool cover -func=coverage.out
-	rm coverage.out
-
-bench:
-	go test -bench=. ./... -run=^$ -v
