+++ leetcode/medium/matrix/lc73_set_matrix_zeroes/73_set_matrix_zeroes.go
@@ -1,4 +1,4 @@
-package matrix
+package lc73_set_matrix_zeroes
 
 /*
 Given an m x n integer matrix matrix, if an element is 0, set its entire row and column to 0's.
